Split commands to separate files
